Replace old webhook with fresh version
